+++ src/pages/AgentsPage.tsx
@@ -1,11 +1,15 @@
 // Copyright (c) 2026 Deus Corp. Licensed under MIT.
 
 import { AgentPanel } from '@/features/agent-panel/AgentPanel'
+import { MetricsStrip } from '@/features/metrics/MetricsStrip'
 
 export function AgentsPage() {
   return (
     <div className="h-full flex flex-col">
-      <AgentPanel />
+      <MetricsStrip />
+      <div className="flex-1 min-h-0">
+        <AgentPanel />
+      </div>
     </div>
   )
 }